requirements.txt
Django==4.0.4
django-recaptcha==3.0.0